deleted basic DCC and ThingWorx
