--- examples/mqtt/dcc_comms/thingworx/sampleProp.conf
@@ -1,6 +1,0 @@
-#MQTT Broker Configurations
-EdgeSystemName = "TestGatewayName"
-BrokerIP = "Broker IP"
-BrokerPort = 1883
-MetricName = "metric-name"
-CustomPubTopic = "liota/gateway/simulated-metric"
