+++ setup.py
@@ -99,9 +99,6 @@ setup(
                 (os.path.abspath(os.sep) + '/../etc/liota/examples/mqtt/device_comms/iot_cc',
                  ['examples/mqtt/device_comms/iot_cc/iotcc_simulated_mqtt.py',
                   'examples/mqtt/device_comms/iot_cc/samplePropMqtt.conf']),
-                (os.path.abspath(os.sep) + '/../etc/liota/examples/mqtt/dcc_comms/thingworx',
-                 ['examples/mqtt/dcc_comms/thingworx/simulated_edge_system.py',
-                  'examples/mqtt/dcc_comms/thingworx/sampleProp.conf']),
                 (os.path.abspath(os.sep) + '/../etc/liota/examples/mqtt/dcc_comms/aws_iot',
                  ['examples/mqtt/dcc_comms/aws_iot/sampleProp.conf',
                   'examples/mqtt/dcc_comms/aws_iot/simulated_home_auto_gen_topic.py',
@@ -136,7 +133,7 @@ setup(
                   'packages/examples//mqtt/simulated_home_topic_per_metric.py']),
                 (os.path.abspath(os.sep) + '/../etc/liota/packages/edge_systems/dell5k',
                  ['packages/edge_systems/dell5k/edge_system.py']),
-		(os.path.abspath(os.sep) + '/../etc/liota/scripts',
+                (os.path.abspath(os.sep) + '/../etc/liota/scripts',
                  ['scripts/autostartliota']),
                 (os.path.abspath(os.sep) + '/../etc/liota/conf',
                  ['config/liota.conf', 'config/logging.json']),
